A few interesting hacks to increase speed:
(1) bucket states so that in multispt, a new state is only added if it
(a) has a different set of trips and (b) is better than an existing
state by at least 30 seconds or 30 weight units and 2%.

(2) DefaultRemainingWeightHeuristic now checks for walk distance
violations, and returns -1, meaning unreachable, when the walk distance
has been exceeded.

(3) In CPSI, when the initial trip has had its walk distance
increased, subsequent trips also have an increased walk distance.